Removed submodules and added flow editor into this repo instead. (#18)
* Removed submodules and added flow editor into this repo instead.

* Fix mac and linux builds.

Trying to fix linux build for spdlog.

Ading position independent code for hello_imgui as well.

Explicitly add fPIC to compile options.

* Updating flow core version.

+++ src/widgets/PropertyTree.cpp
@@ -65,7 +65,8 @@ void PropertyTree::operator()() noexcept
                 continue;
             }
 
-            auto property_table = Table(category_name + "##table", _columns, ImGui::GetItemRectSize().x + 1.f);
+            auto property_table =
+                Table(category_name + "##table", _columns, static_cast<std::size_t>(ImGui::GetItemRectSize().x + 1.f));
             for (const auto& widget : widgets)
             {
                 property_table.AddEntry(widget);
